+++ frontend/src/pages/CreatePage/CreatePage.tsx
@@ -1,47 +1,17 @@
 import Text from '~/components/common/Text/Text';
 import * as S from './CreatePage.styled';
 import Input from '~/components/common/Input/Input';
-import {
-  type ChangeEventHandler,
-  type FormEventHandler,
-  useState,
-  useRef,
-} from 'react';
+import { useRef } from 'react';
 import Button from '~/components/common/Button/Button';
-import { useSendNewTeamPlace } from '~/hooks/queries/useSendNewTeamPlace';
-import { useNavigate } from 'react-router-dom';
-import { PATH_NAME } from '~/constants/routes';
-import { KEY } from '~/constants/localStorage';
+import { useTeamCreate } from '~/hooks/team/useTeamCreate';
 
 const CreatePage = () => {
-  const [teamName, setTeamName] = useState('');
   const inputRef = useRef<HTMLInputElement>(null);
-  const navigate = useNavigate();
-  const { mutateSendNewTeamPlace } = useSendNewTeamPlace();
+  const {
+    teamName,
 
-  const handleTeamNameChange: ChangeEventHandler<HTMLInputElement> = (e) => {
-    setTeamName(() => e.target.value);
-  };
-
-  const handleTeamNameSubmit: FormEventHandler<HTMLFormElement> = (e) => {
-    e.preventDefault();
-
-    const isRightName = confirm(`"${teamName}"으로 팀을 생성하시겠습니까?`);
-
-    if (!isRightName) {
-      inputRef.current?.focus();
-      return;
-    }
-    mutateSendNewTeamPlace(
-      { name: teamName },
-      {
-        onSuccess: (data) => {
-          localStorage.setItem(KEY.TEAM_PLACE_ID, String(data.teamPlaceId));
-          navigate(PATH_NAME.TEAM_SELECT);
-        },
-      },
-    );
-  };
+    handler: { handleTeamNameChange, handleTeamNameSubmit },
+  } = useTeamCreate(inputRef);
 
   return (
     <S.Container>
